Extract UnusedParameter
